adds prisma schema and prisma adapter for next-auth
