+++ index.d.ts
@@ -1,0 +1,7 @@
+import type { PrismaClient } from '@prisma/client'
+
+declare global {
+    namespace globalThis {
+        var prisma: PrismaClient;
+    }
+}
