--- pages/api/auth/[...nextauth].js
@@ -1,16 +1,0 @@
-import NextAuth from "next-auth";
-import GithubProvider from "next-auth/providers/github";
-
-export const authOptions = {
-  // Configure one or more authentication providers
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID,
-      clientSecret: process.env.GITHUB_SECRET,
-    }),
-    // ...add more providers here
-  ],
-  secret: process.env.JWT_SECRET,
-};
-
-export default NextAuth(authOptions);
